fixed some html errors Removed the need to refresh the tree when changing the print options while in printing mode integrated trackplus support Added checking for duplicate testproject names added printing of testplans added initial sql-scripts for mssql for mssql the installer has to write something like this 	define('DB_TYPE', 'odbc_mssql'); 	define('DB_USER', 'user'); 	define('DB_PASS', 'pass'); 	define('DB_HOST', 'Driver={SQL Server};Server=SERVER\SQLEXPRESS;Database=testlink;'); 	define('DB_NAME', 'testlink');
slightly changed the layout of the mainpage
removed some warning in attachment related code (PHP5)
changed the definition of users.role_id from
	`role_id` tinyint(3) unsigned NOT NULL default '0',
  to
	`role_id` int(10) unsigned NOT NULL default '0',

corrected get_by_name in testproject.class.php
corrected new_node in tree.class.php

+++ install/sql/mysql/testlink_create_tables.sql
@@ -1,6 +1,6 @@
 # TestLink Open Source Project - http://testlink.sourceforge.net/
 # This script is distributed under the GNU General Public License 2 or later.
-# $Id: testlink_create_tables.sql,v 1.4 2006/07/18 06:39:11 franciscom Exp $
+# $Id: testlink_create_tables.sql,v 1.5 2006/07/28 17:22:04 schlundus Exp $
 # SQL script - create db tables for TL   
 #
 # default rights & admin account are created via testlink_create_default_data.sql
@@ -256,7 +256,7 @@ CREATE TABLE `users` (
   `id` int(10) unsigned NOT NULL auto_increment,
   `login` varchar(30) NOT NULL default '',
   `password` varchar(32) NOT NULL default '',
-  `role_id` tinyint(3) unsigned NOT NULL default '0',
+  `role_id` int(10) unsigned NOT NULL default '0',
   `email` varchar(100) NOT NULL default '',
   `first` varchar(30) NOT NULL default '',
   `last` varchar(30) NOT NULL default '',
